Issue #74: Ensure unique search results. Extracted sitemap to a dedicated module.
