+++ tales_django/core/pages/get_tracks_list_context.py
@@ -35,7 +35,7 @@ def get_tracks_list_context(request: HttpRequest):
     filter_kwargs = get_track_filter_kwargs(request)
     filter_args = get_search_filter_args(request)
 
-    tracks = Track.objects.filter(*filter_args, **filter_kwargs).order_by(*order_args)
+    tracks = Track.objects.filter(*filter_args, **filter_kwargs).distinct().order_by(*order_args)
     # tracks = Track.objects.filter(track_status='PUBLISHED').order_by('-published_at', f'title_{language}').all()
 
     tracks_offset = int(request.GET.get('tracks_offset', 0))
